requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=0.21.1
pydantic==1.10.7
jinja2>=3.1.2
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.7.0
aiosmtplib>=2.0.1
email-validator>=1.1.0,<2.0.0
bcrypt>=4.0.1
itsdangerous>=2.1.2
PyJWT>=2.8.0
python-slugify>=8.0.1
redis>=5.0.1
hiredis>=2.3.2
